upd: Refactor and make it possible to configure multiple directories and namespaces
