+++ composer.json
@@ -14,6 +14,6 @@
     ],
     "license": "MIT",
     "require": {
-        "phpstan/phpstan": "^1.8.1"
+        "phpstan/phpstan": "^1.10"
     }
 }
